isStartAndEndNodesValid now uses the parameter for number of nodes instead of checking if the node is available in the collection. The case of the user asking for a node without neighbours as a start or ending node will be handles as a special case by the GetShortestPathBetweenNodes method, so we dont need to create a lot of unneded nodes.
Changed the property Neighbours on the Node class to a struct that holds both Neighbour node ID as well as a weight of the path

Modified the Node constructor to handle the Neighbour struct

Removed a test temporarily becase I could not wrap my head around the linq statement that needed to be changed because of the new neighbours struct. That will be revisited later today. Lunchtime!
Added a Node constructor to create nodes with no neighbours

Modified Dijkstra so it will use path weights

Modified GetShortestPathBetweenNodes so it will handle the case of a user entering a node with no paths as a start or end value.